+++ pkg/contractManager/contractManager_test.go
@@ -2,6 +2,7 @@ package contractManager
 
 import (
 	"context"
+	"database/sql"
 	"log"
 	"testing"
 
@@ -12,9 +13,13 @@ import (
 	"github.com/Layr-Labs/sidecar/internal/metrics"
 	"github.com/Layr-Labs/sidecar/internal/tests"
 	"github.com/Layr-Labs/sidecar/pkg/clients/ethereum"
+	"github.com/Layr-Labs/sidecar/pkg/contractStore"
 	"github.com/Layr-Labs/sidecar/pkg/contractStore/postgresContractStore"
+	"github.com/Layr-Labs/sidecar/pkg/parser"
 	"github.com/Layr-Labs/sidecar/pkg/postgres"
+	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
+	"github.com/ethereum/go-ethereum/common"
 	"gorm.io/gorm"
 )
 
@@ -41,13 +46,14 @@ func setup() (
 }
 
 func Test_ContractManager(t *testing.T) {
-	_, grm, l, cfg, err := setup()
+	// uses setup from restakedStrategies_test
+	dbName, grm, l, cfg, err := setup()
 
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	baseUrl := "https://tame-fabled-liquid.quiknode.pro/f27d4be93b4d7de3679f5c5ae881233f857407a0/"
+	baseUrl := "https://winter-white-crater.ethereum-holesky.quiknode.pro/1b1d75c4ada73b7ad98e1488880649d4ea637733/"
 	ethConfig := ethereum.DefaultNativeCallEthereumClientConfig()
 	ethConfig.BaseUrl = baseUrl
 
@@ -56,6 +62,19 @@ func Test_ContractManager(t *testing.T) {
 	metricsClients, err := metrics.InitMetricsSinksFromConfig(cfg, l)
 	if err != nil {
 		l.Sugar().Fatal("Failed to setup metrics sink", zap.Error(err))
+	}
+
+	contract := &contractStore.Contract{
+		ContractAddress:         "0x123",
+		ContractAbi:             "[]",
+		Verified:                true,
+		BytecodeHash:            "0x123",
+		MatchingContractAddress: "",
+	}
+	proxyContract := &contractStore.ProxyContract{
+		BlockNumber:          1,
+		ContractAddress:      contract.ContractAddress,
+		ProxyContractAddress: "0x456",
 	}
 
 	sdc, err := metrics.NewMetricsSink(&metrics.MetricsSinkConfig{}, metricsClients)
@@ -68,8 +87,63 @@ func Test_ContractManager(t *testing.T) {
 		log.Fatalf("Failed to initialize core contracts: %v", err)
 	}
 
-	t.Run("Test fetching ABI from IPFS", func(t *testing.T) {
+	t.Run("Test indexing contract upgrades", func(t *testing.T) {
+		// Create a contract
+		_, err := contractStore.CreateContract(contract.ContractAddress, contract.ContractAbi, contract.Verified, contract.BytecodeHash, contract.MatchingContractAddress, false)
+		assert.Nil(t, err)
+
+		// Create a proxy contract
+		_, err = contractStore.CreateProxyContract(uint64(proxyContract.BlockNumber), proxyContract.ContractAddress, proxyContract.ProxyContractAddress)
+		assert.Nil(t, err)
+
+		// Check if contract and proxy contract exist
+		var contractCount int
+		contractAddress := contract.ContractAddress
+		res := grm.Raw(`select count(*) from contracts where contract_address=@contractAddress`, sql.Named("contractAddress", contractAddress)).Scan(&contractCount)		
+		assert.Nil(t, res.Error)
+		assert.Equal(t, 1, contractCount)
+
+		proxyContractAddress := proxyContract.ContractAddress
+		res = grm.Raw(`select count(*) from contracts where contract_address=@proxyContractAddress`, sql.Named("proxyContractAddress", proxyContractAddress)).Scan(&contractCount)		
+		assert.Nil(t, res.Error)
+		assert.Equal(t, 1, contractCount)
+
+		var proxyContractCount int
+		res = grm.Raw(`select count(*) from proxy_contracts where contract_address=@contractAddress`, sql.Named("contractAddress", contractAddress)).Scan(&proxyContractCount)		
+		assert.Nil(t, res.Error)
+		assert.Equal(t, 1, proxyContractCount)
+
+		// An upgrade event
+		upgradedLog := &parser.DecodedLog{
+			LogIndex:  0,
+			Address:   contract.ContractAddress,
+			EventName: "Upgraded",
+			Arguments: []parser.Argument{
+				{
+					Name:    "implementation",
+					Type:    "address",
+					Value:   common.HexToAddress("0x789"),
+					Indexed: true,
+				},
+			},
+		}
+
+		// Perform the upgrade
 		cm := NewContractManager(contractStore, client, sdc, l)
-		_ = cm.FetchAbiFromIPFS(context.Background(), "0xdabdb3cd346b7d5f5779b0b614ede1cc9dcba5b7")
+		err = cm.HandleContractUpgrade(context.Background(), 5, upgradedLog)
+		assert.Nil(t, err)
+
+		// Verify database state after upgrade
+		newProxyContractAddress := upgradedLog.Arguments[0].Value.(common.Address).Hex()
+		res = grm.Raw(`select count(*) from contracts where contract_address=@newProxyContractAddress`, sql.Named("newProxyContractAddress", newProxyContractAddress)).Scan(&contractCount)
+		assert.Nil(t, res.Error)
+		assert.Equal(t, 1, contractCount)
+
+		res = grm.Raw(`select count(*) from proxy_contracts where contract_address=@contractAddress`, sql.Named("contractAddress", contractAddress)).Scan(&proxyContractCount)		
+		assert.Nil(t, res.Error)
+		assert.Equal(t, 2, proxyContractCount)
+	})	
+	t.Cleanup(func() {
+		postgres.TeardownTestDatabase(dbName, cfg, grm, l)
 	})
 }
