+++ pkg/fetcher/fetcher.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"github.com/Layr-Labs/sidecar/internal/config"
 	"github.com/Layr-Labs/sidecar/pkg/clients/ethereum"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 	"slices"
@@ -243,10 +242,4 @@ func (f *Fetcher) FetchBlocks(ctx context.Context, startBlockInclusive uint64, e
 	)
 
 	return fetchedBlocks, nil
-}
-
-func (f *Fetcher) GetContractStorageSlot(ctx context.Context, contractAddress string, blockNumber uint64) (string, error) {
-	stringBlock := hexutil.EncodeUint64(blockNumber)
-	
-	return f.EthClient.GetStorageAt(ctx, contractAddress, ethereum.EIP1967_STORAGE_SLOT, stringBlock)
 }
